feat: id로 키워드 조회

+++ src/main/java/baby/ey/diary/repository/GrowthRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface GrowthRepository extends JpaRepository<Growth, Integer> {
+public interface GrowthRepository extends JpaRepository<Growth, Long> {
     List<Growth> findAllByOrderByCreatedDesc();
 }
